+++ level.go
@@ -3,6 +3,7 @@ package goseq
 // Level represents the log level
 type Level int
 
+//Log level supported by Seq
 const (
 	VERBOSE Level = iota
 	DEBUG