+++ seq_client.go
@@ -3,7 +3,7 @@ package goseq
 import (
 	"bytes"
 	"encoding/json"
-	"fmt"
+	"errors"
 	"log"
 	"net/http"
 )
@@ -18,7 +18,7 @@ type SeqClient struct {
 }
 
 // Send send POST requests to the SEQ API
-func (sc *SeqClient) Send(event *SeqLog, apiKey string, client *http.Client) bool {
+func (sc *SeqClient) Send(event *SeqLog, apiKey string, client *http.Client) error {
 
 	fullURL := sc.BaseURL + endpoint
 
@@ -37,13 +37,13 @@ func (sc *SeqClient) Send(event *SeqLog, apiKey string, client *http.Client) boo
 	response, err := client.Do(request)
 	defer request.Body.Close()
 	if err != nil {
-		fmt.Println(err)
-		return false
+
+		return err
 	}
 	defer response.Body.Close()
 
-	if response.StatusCode == 201 {
-		return true
+	if response.StatusCode != 201 {
+		return errors.New(response.Status)
 	}
-	return false
+	return nil
 }