+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 Added
 - Adds support to n (n > 0) http Seq consumer at Logger to prevent the application to consume a lot of memory when we have a thousands of requests
-- Change SeqClient return data, from bool to error
+- Change SeqClient.Send return type, from bool to error
 
 ## 0.1.1 (2017-03-17)